# Optimizations: do less redundant work, part 2