… src/js/dom/index.js
… src/js/dom/refs.js
… src/js/dom/tree-walker.js
… src/js/event.js
… src/js/form-model.js
… src/js/form.js
… src/js/input.js
… src/js/itemset.js
… src/js/language.js
… src/js/output.js
… src/js/page.js
… src/js/progress.js
… src/js/readonly.js
… src/js/relevant.js
… src/js/repeat.js
… src/js/required.js
… src/js/widget.js
… src/js/widgets-controller.js
… src/widget/time/timepicker.js
… test/forms/collections.xml
… test/spec/calculate.spec.js
… test/spec/dom/collections.spec.js
… test/spec/dom/tree-walker.spec.js
… test/spec/input.spec.js
… test/spec/itemset.spec.js
… test/spec/repeat.spec.js
… tsconfig.json