@lognaturel 2023-05-23
@eyelidlessness and I did some interactive review. I have thought through each class of change and have a high-level understanding of each.

We are going to merge and update in Express master so it gets exercised on https://getodk.org/xlsform/. We'll also do a QA pass. Feedback and testing by other users is very welcome.